requirements.txt
Django==1.5.2
South==0.8.2
argparse==1.2.1
boto==2.13.3
distribute==0.6.34
dj-database-url==0.2.2
dj-static==0.0.5
django-autoslug==1.7.1
django-cors-headers==0.12
django-filter==0.7
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==2.3.7
gunicorn==18.0
ipython==1.0.0
newrelic==2.0.0.1
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
